build: produce CommonJS module only
Evidently, producing both a CommonJS and ES module using TypeScript from the
same source is not easy, at least with the current project structure (which uses
two separate output files). At this time, using CommonJS only seems to be the
best solution for interoperability, but this may be revisited later.