build: remove transpilation steps and move to ES modules
This commit removes the need for TypeScript as a transpiler, preferring to use
JSDoc comments and plain JS files and generate type declarations from them. This
makes the build setup a little cleaner and less tedious (no need to rebuild
things when testing locally with file modules in package.json, etc.).

As an intentional side-effect, this commit also migrates the project to be ESM
only.

BREAKING CHANGE: the library is no longer distributed as a CommonJS module.