build: move everything into index.js
This commit moves the entire library, including the automatic jsx-runtime, into
index.js. This relies on the `exports` field in package.json to handle mapping
imports of `static-jsx/jsx-runtime` back to index.js, which means that tools
such as TypeScript which don't understand this will not work with the automatic
transform. However, this is much easier to maintain than having a separate
jsx-runtime directory, which caused a lot of issues for various tools.

BREAKING CHANGE: the automatic JSX transform now implicitly depends on support
for the `exports` field in package.json.